# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Smart File Backup System - Team D Master API
  description: Master OpenAPI contract for Team D Web Frontend & Backend-For-Frontend (BFF), orchestrating Teams A, B, and C.
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/v1
    description: Team D Local BFF Server
paths:
...
  /ui/files/{fileId}:
    patch:
      summary: Update file retention policy or storage location (IT Admin)
      operationId: updateFilePolicy
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                retention_days:
                  type: integer
                  example: 90
                storage_class:
                  type: string
                  example: "COLD_STORAGE"
      responses:
        '200':
          description: File policy updated
  /ui/backups:
    post:
      summary: Schedule a file backup operation
      operationId: startBackup
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BackupRequest'
      responses:
        '202':
          description: Backup request accepted and queued
  /ui/backups/{backupId}:
    get:
      summary: Check status and state machine progress of a backup
      operationId: getBackupStatus
      parameters:
        - name: backupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed backup state, worker lease, dedup ratio, and Merkle verification
...
components:
  schemas:
    BackupRequest:
      type: object
      required: [backup_id, user_id, file_id, backup_type, source_path, priority, idempotency_key]
      properties:
        backup_id:
          type: string
        user_id:
          type: string
        file_id:
          type: string
        backup_type:
          type: string
          enum: [FULL, DIFFERENTIAL, INCREMENTAL]
        source_path:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 5
        idempotency_key:
          type: string
...